Master

+++ install.sh
@@ -36,6 +36,10 @@ workdir=$(pwd)
 
 #Install Basic Tools
 if [ ! -e /usr/local/bin/ssr ];then
+if [[ $1 == "uninstall" ]];then
+	echo "你在开玩笑吗？你都没有安装怎么卸载呀！"
+	exit 1
+fi
 echo "开始部署"
 sleep 2s
 if [[ ${OS} == Ubuntu ]];then
@@ -44,12 +48,15 @@ if [[ ${OS} == Ubuntu ]];then
 	apt-get install python-pip -y
 	apt-get install git -y
 	apt-get install language-pack-zh-hans -y
+	apt-get -y install vnstat bc
     apt-get install build-essential screen curl -y
 fi
 if [[ ${OS} == CentOS ]];then
 	yum install python screen curl -y
 	yum install python-setuptools -y && easy_install pip -y
 	yum install git -y
+	yum install bc -y
+	yum install vnstat
     yum groupinstall "Development Tools" -y
 fi
 if [[ ${OS} == Debian ]];then
@@ -57,23 +64,36 @@ if [[ ${OS} == Debian ]];then
 	apt-get install python screen curl -y
 	apt-get install python-pip -y
 	apt-get install git -y
+	apt-get -y install bc vnstat
     apt-get install build-essential -y
 fi
 #Install Libsodium
-cd $workdir
-export LIBSODIUM_VER=1.0.13
-wget https://github.com/jedisct1/libsodium/releases/download/1.0.13/libsodium-$LIBSODIUM_VER.tar.gz
-tar xvf libsodium-$LIBSODIUM_VER.tar.gz
-pushd libsodium-$LIBSODIUM_VER
-./configure --prefix=/usr && make
-make install
-popd
-ldconfig
-cd $workdir && rm -rf libsodium-$LIBSODIUM_VER.tar.gz libsodium-$LIBSODIUM_VER
+libsodiumfilea="/usr/local/lib/libsodium.so"
+libsodiumfileb="/usr/lib/libsodium.so"
+if [[ -e ${libsodiumfilea} ]];then
+    echo "libsodium已安装!"
+elif [[ -e ${libsodiumfileb} ]];then
+    echo "libsodium已安装!"
+else
+    cd $workdir
+    export LIBSODIUM_VER=1.0.15
+    wget -q https://github.com/jedisct1/libsodium/releases/download/${LIBSODIUM_VER}/libsodium-$LIBSODIUM_VER.tar.gz
+    tar xvf libsodium-$LIBSODIUM_VER.tar.gz
+    pushd libsodium-$LIBSODIUM_VER
+    ./configure --prefix=/usr && make
+    make install
+    popd
+    ldconfig
+    cd $workdir && rm -rf libsodium-$LIBSODIUM_VER.tar.gz libsodium-$LIBSODIUM_VER
+#    if [[ ! -e ${libsodiumfile} ]];then
+#    	echo "libsodium安装失败 !"
+#    	exit 1
+#    fi
+fi
 cd /usr/local
 git clone https://github.com/Readour/shadowsocksr.git
 cd ./shadowsocksr
-git manyuser
+git checkout manyuser
 git pull
 if [ $1 == "develop" ];then
     git checkout stack/dev
@@ -82,11 +102,44 @@ fi
 
 #Install SSR and SSR-Bash
 if [ -e /usr/local/bin/ssr ];then
+	if [[ $1 == "uninstall" ]];then
+		echo "开始卸载"
+		sleep 1s
+		echo "删除:/usr/local/bin/ssr"
+		rm -f /usr/local/bin/ssr
+		echo "删除:/usr/local/SSR-Bash-Python"
+		rm -rf /usr/local/SSR-Bash-Python
+		echo "删除:/usr/local/shadowsocksr"
+		rm -rf /usr/local/shadowsocksr
+		echo "删除:${PWD}/install.sh"
+		rm -f ${PWD}/install.sh
+		sleep 1s
+		echo "卸载完成!!"
+		exit 0
+	fi
+    if [[ ! $yn == n ]];then
+        echo "你是否为其它版本迁移而来？（Y/N）"
+        read -t 2 -n 1 yn
+    fi
+    if [[ ${yn} == [yY] ]];then
+        mv /usr/local/shadowsocksr/mudb.json /usr/local/mudb.json
+        rm -rf /usr/local/shadowsocksr
+        cd /usr/local
+        git clone https://github.com/Readour/shadowsocksr.git
+        if [[ $1 == develop ]];then
+            cd ./shadowsocksr
+            git checkout stack/dev
+            rm -f ./mudb.json
+            mv ../mudb.json ./mudb.json
+        else
+            rm -f ./shadowsocksr/mudb.json
+            mv /usr/local/mudb.json /usr/local/shadowsocksr/mudb.json
+        fi
+    fi
 	echo "开始更新"
 	sleep 1s
 	echo "正在清理老版本"
 	rm -f /usr/local/bin/ssr
-	rm -rf /usr/local/SSR-Bash-Python
 	sleep 1s
 	echo "开始部署"
 	cd /usr/local/shadowsocksr
@@ -96,18 +149,41 @@ if [ -e /usr/local/bin/ssr ];then
         git checkout stack/dev
     fi
 fi
-cd /usr/local
-git clone https://github.com/Readour/AR-B-P-B.git
-cd AR-B-P-B
-git checkout master
-if [[ $1 == "develop" ]];then
-    git checkout develop
+if [[ -d /usr/local/SSR-Bash-Python ]];then
+    if [[ $yn == [yY] ]];then
+        rm -rf /usr/local/SSR-Bash-Python
+        cd /usr/local
+        git clone https://github.com/Readour/AR-B-P-B.git
+        mv AR-B-P-B SSR-Bash-Python
+    fi
+    cd /usr/local/SSR-Bash-Python
+    git checkout master
+    git pull
+    if [[ $1 == "develop" ]];then
+        git checkout develop
+        git pull
+    fi
+else
+    cd /usr/local
+    git clone https://github.com/Readour/AR-B-P-B.git
+    cd AR-B-P-B
+    git checkout master
+    if [[ $1 == "develop" ]];then
+        git checkout develop
+    fi
+    cd ..
+    mv AR-B-P-B SSR-Bash-Python
 fi
-cd ..
-mv AR-B-P-B SSR-Bash-Python
 cd /usr/local/shadowsocksr
 bash initcfg.sh
-
+if [[ ! -e /usr/bin/bc ]];then
+	if [[ ${OS} == CentOS ]];then
+		yum install bc -y
+	fi
+	if [[ ${OS} == Ubuntu || ${OS} == Debian ]];then
+		apt-get install bc -y
+	fi
+fi
 #Start when boot
 if [[ ${OS} == Ubuntu || ${OS} == Debian ]];then
     cat >/etc/init.d/ssr-bash-python <<EOF
@@ -170,20 +246,52 @@ systemctl enable iptables.service
 fi
 
 #Install SSR-Bash Background
-wget -N --no-check-certificate -O /usr/local/bin/ssr https://raw.githubusercontent.com/readour/AR-B-P-B/master/ssr
-chmod +x /usr/local/bin/ssr
+if [[ $1 == "develop" ]];then
+	wget -q -N --no-check-certificate -O /usr/local/bin/ssr https://raw.githubusercontent.com/readour/AR-B-P-B/develop/ssr
+	chmod +x /usr/local/bin/ssr
+else
+	wget -q -N --no-check-certificate -O /usr/local/bin/ssr https://raw.githubusercontent.com/readour/AR-B-P-B/master/ssr
+	chmod +x /usr/local/bin/ssr
+fi
 
 #Modify ShadowsocksR API
 sed -i "s/sspanelv2/mudbjson/g" /usr/local/shadowsocksr/userapiconfig.py
 sed -i "s/UPDATE_TIME = 60/UPDATE_TIME = 10/g" /usr/local/shadowsocksr/userapiconfig.py
 sed -i "s/SERVER_PUB_ADDR = '127.0.0.1'/SERVER_PUB_ADDR = '$(wget -qO- -t1 -T2 ipinfo.io/ip)'/" /usr/local/shadowsocksr/userapiconfig.py
 #INstall Success
-read -p "输入与您主机绑定的域名: " ipname
+read -t 15 -p "输入与您主机绑定的域名(请在15秒内输入，超时将跳过本步骤): " ipname
 echo "$ipname" > /usr/local/shadowsocksr/myip.txt
+if [[ $1 == develop ]];then
+    if [[ -e /usr/local/SSR-Bash-Python/check.log ]];then
+        cd /usr/local/SSR-Bash-Python
+        read -n 1 -t 3 -p "你是否想要重新配置服务器巡检配置（注意，这将会清空你的日志）[Y/N]" yn
+        if [[ $yn == [Yy] ]];then
+        	bash servercheck.sh reconf
+        	nohup bash servercheck.sh run 2>/dev/null &
+        else
+        	bash servercheck.sh stop
+        	nohup bash servercheck.sh run 2>/dev/null &
+            echo "服务已重启"
+        fi
+    else
+        read -t 10 -p "是否设置服务器自检，实验型功能！[Y/N]" yn
+        if [[ $yn == [yY] ]];then
+        	cd /usr/local/SSR-Bash-Python
+        	bash servercheck.sh conf
+        	nohup bash servercheck.sh run 2>/dev/null &
+        	PID=$(ps -ef |grep -v grep | grep "bash" | grep "servercheck.sh" | grep "run" | awk '{print $2}')
+        	if [[ -z ${PID} ]];then
+            	echo "程序启动失败,请联系作者"
+            fi
+        else
+        	echo "你居然拒绝了T.T"
+        fi
+    fi
+fi
 bash /usr/local/SSR-Bash-Python/self-check.sh
 echo '安装完成！输入 ssr 即可使用本程序~'
-echo '原作者已经停止本脚本服务，此版本为2017.7.30号的备份（带最新端口限速）'
-echo '还好在作者删服务前备份了'
+echo '原作者已经停止本脚本更新，此版本为作者删除项目前最后一个版本魔改而来'
+echo '不喜勿喷!'
 echo '谨慎使用！仅供研究！'
 echo '谨慎使用！仅供研究！'
 echo '谨慎使用！仅供研究！'
